requirements.txt
astroid==2.3.3
certifi==2019.11.28
Django==2.2.8
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
PyMySQL==0.9.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wrapt==1.11.2
